# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Punjabi Rishtey Matrimony API",
    "version": "1.0.0",
    "description": "API documentation for user and admin panel"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local server"
    }
  ],
  "paths": {
    "/api/admin/auth/register": {
      "post": {
        "summary": "Register a new admin",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Admin registered successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/admin/auth/login": {
      "post": {
        "summary": "Admin login",
        "tags": [
          "Admin"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Login successful"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/dashboard": {
      "get": {
        "summary": "Get admin dashboard statistics",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Admin dashboard data retrieved successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/users/{status}": {
      "get": {
        "summary": "Get users by status",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "status",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The status of users (approved, pending, expired, canceled)"
          }
        ],
        "responses": {
          "200": {
            "description": "Users fetched successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/users/approve/{id}": {
      "put": {
        "summary": "Approve a user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to approve"
          }
        ],
        "responses": {
          "200": {
            "description": "User approved successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/users/block/{id}": {
      "put": {
        "summary": "Block a user",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to block"
          }
        ],
        "responses": {
          "200": {
            "description": "User blocked successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/users/edit/{id}": {
      "put": {
        "summary": "Edit user details",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The ID of the user to edit"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User details updated successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/users/add": {
      "post": {
        "summary": "Add a new user from admin panel",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "password": {
                    "type": "string"
                  },
                  "status": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User added successfully"
          },
          "400": {
            "description": "Bad request"
          }
        }
      }
    },
    "/api/admin/auth/dashboard/registrations": {
      "get": {
        "summary": "Get monthly user registrations data",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Monthly registration data retrieved successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/admin/auth/dashboard/status-counts": {
      "get": {
        "summary": "Get user status counts",
        "tags": [
          "Admin"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User status counts retrieved successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
    "/api/users/{id}": {
      "get": {
        "summary": "Get user profile by ID",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "put": {
        "summary": "Update user profile",
        "tags": [
          "Users"
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string"
                  },
                  "gender": {
                    "type": "string"
                  },
                  "dob": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully"
          },
          "403": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}